Fix candidate reload bug, organize widget files
