+++ src/js/components/Widgets/ItemSupportOpposeCounts.jsx
