--- src/js/mock-data/data.js
@@ -1,107 +1,0 @@
-module.exports = {
-    ballot : {
-        status: "VOTER_BALLOT_ITEMS_RETRIEVED",
-        "google_civic_election_id": "2000",
-        "ballot_item_list": [
-            {
-                "we_vote_id": "wvteoff1",
-                "type": "office",
-                "local_ballot_order":1,
-                "google_civic_election_id": "2000",
-                "google_ballot_placement":0,
-                "id": "444",
-                "ballot_item_label": "US Senate"
-
-            },{
-                "we_vote_id": "wvteoff2",
-                "type": "office",
-                "local_ballot_order":2,
-                "google_civic_election_id": "2000",
-                "google_ballot_placement":0,
-                "id": "445",
-                "ballot_item_label": "US House - District 8"
-
-            },{
-                "we_vote_id": "wvteoff3",
-                "type": "office",
-                "local_ballot_order": 3,
-                "google_civic_election_id": "2000",
-                "google_ballot_placement":0,
-                "id": "446",
-                "ballot_item_label": "Treasurer"
-
-            },{
-                "we_vote_id": "wvteoff4",
-                "type": "office",
-                "local_ballot_order":4,
-                "google_civic_election_id": "2000",
-                "google_ballot_placement":0,
-                "id": "447",
-                "ballot_item_label": "Member School Board - Special"
-
-            },{
-                "we_vote_id": "wvteoff5",
-                "type": "office",
-                "local_ballot_order":5,
-                "google_civic_election_id": "2000",
-                "google_ballot_placement":0,
-                "id": "448",
-                "ballot_item_label": "Member School Board"
-
-            },{
-                "we_vote_id": "wvteoff6",
-                "type": "office",
-                "local_ballot_order":6,
-                "google_civic_election_id": "2000",
-                "google_ballot_placement":0,
-                "id": "449",
-                "ballot_item_label": "Member County Board"
-
-            },{
-                "we_vote_id": "wvteoff7",
-                "type": "office",
-                "local_ballot_order":7,
-                "google_civic_election_id": "2000",
-                "google_ballot_placement":8,
-                "id": "450",
-                "ballot_item_label": "Member House of Delegates"
-            }
-        ],
-        "voter_device_id": "Zish7FeoMrOR2tgkPlSZ5vVL3C6CiXyOTDyX7SdfRZdfdUD0VQhYd3UiyUDCBo4pC5i6xz14E1oZsNV1KKPnXq7B",
-        "success": true
-    },
-
-    candidates: [
-        {
-            "office_id": "444",
-            "candidate_list": [
-                {
-                    "candidate_id": 734,
-                    "candidate_display_name": "Ed W. Gillespie",
-                    "order_on_ballot": "0",
-                    "candidate_we_vote_id": "wvtecand1",
-                    "candidate_photo_url": ""
-
-                },{
-
-                    "candidate_id": 735,
-                    "candidate_display_name": "Mark R. Warner",
-                    "order_on_ballot": "0",
-                    "candidate_we_vote_id": "wvtecand2",
-                    "candidate_photo_url": ""
-
-                },{
-                    "candidate_id": 736,
-                    "candidate_display_name": "Robert C. Sarvis",
-                    "order_on_ballot": "0",
-                    "candidate_we_vote_id": "wvtecand3",
-                    "candidate_photo_url": ""
-                }
-            ],
-            "status": "CANDIDATES_RETRIEVED",
-            "google_civic_election_id": "2000",
-            "office_we_vote_id": "wvteoff1",
-            "success": true
-        },
-    ]
-};
